Add another include directory to make it compile on other platforms
This problem is seen on Ubuntu in particular where another include
directory is needed to get C libraries to compile.

I don't know which platforms have this problem or if there are other
platforms where different directories are required. For now this will
work since Kotlin Native is in beta so things may change