Suppress warning when constructing the position values
The position object is created in MathCombineAPLFunction based on the name of function itself. The name is retrieved using the name1Arg and name2Arg functions. A warning is raised when reading the name, as it calls a subclass while the parent is being initialised.

This call is safe, since the actual implementations of these functions do not rely on the state of the function itself.

It should be noted that there may be cases where this call is no longer safe, but there are no such cases at the moment.